Set `updateDelegate` when a provider invalidates
Basic `Section`s were not propagating updates if added via a call to `invalidateAll`. This is to address #2.

The issue can be recreated by using a `SegmentedSectionProvider` when switching to a segment that contains a `Section` (`SectionProvider` works I _think_)

This commit alone is likely not enough to fix this issue; `updateDelegate` should probably be set to `nil` for removed `Section`s, and should probably be set in more places.