Fix tests, don’t pass ChangesReducer to `updates`
The caller does not need to worry about how changes are batched and this will allow other UI implementations to have their batching mechanisms (if required, or not at all if not required).

This also moves `ChangesReducer` back to `ComposedUI` since it’s really only applicable to the UI part.