Correctly calculate `CollectionFlowLayoutEnvironment.contentSize`
The layout's `sectionInset` property is applied on a per-section basis, but the collection view's `contentInset` was not being honoured.

It was also calculated using `insetBy(dx:dy:)`, which will modify the width by `-dx * 2`, which would double the expected insets.

This is technically a breaking change for those relying on this behaviour, but it's a big enough bug that it should probably still be released without a major version bump. (plus not many are using the porject anyway 😬)